more tests for judging process:
