+++ package.json
@@ -29,6 +29,7 @@
     "chai-as-promised": "^7.1.1",
     "eth-hd-wallet": "^0.5.1",
     "ethereum-event-logs": "^1.3.1",
+    "ethval": "^2.0.0",
     "ganache-cli": "^6.8.2",
     "js-sha3": "^0.8.0",
     "lodash": "^4.17.15",
